Extract auth module: move JWT, OAuth2, security config, auth endpoints, password management from application/user to auth
- New auth module (pt.estga.auth): SecurityConfig, JwtService, RsaKeyConfig, AppJwtAuthenticationConverter, AuthApplicationConfig, OpenApiConfig, AuditingConfig
- OAuth: GoogleOAuth2UserService, OAuth2LoginSuccessHandler, GoogleTokenVerifier
- Controllers: AuthController, AccountPasswordController (extracted from user AccountController)
- Services: AuthService, PasswordService (extracted from user UserService)
- DTOs: AuthResponse, LoginRequest, GoogleAuthRequest, RefreshTokenRequest, PasswordSetRequest, PasswordChangeRequest
- Resources: application-security.yml
- Application module: removed security starter dependencies (transitive from auth), simplified ApplicationConfig
- User module: removed setPassword/changePassword from UserService, removed password endpoints from AccountController
- Root pom: added auth module, dependency management
