+++ application/pom.xml
@@ -92,16 +92,9 @@
             <version>${stonemark.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-oauth2-client</artifactId>
+            <groupId>pt.estga</groupId>
+            <artifactId>auth</artifactId>
+            <version>${stonemark.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
