+++ auth/src/main/java/pt/estga/auth/config/JwtService.java
@@ -1,4 +1,4 @@
-package pt.estga.config;
+package pt.estga.auth.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.jwt.Jwt;
