+++ auth/src/main/java/pt/estga/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package pt.estga.dtos;
+package pt.estga.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
