+++ auth/src/main/java/pt/estga/auth/dto/PasswordChangeRequest.java
@@ -1,0 +1,9 @@
+package pt.estga.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record PasswordChangeRequest(
+        @NotBlank String oldPassword,
+        @NotBlank @Size(min = 8, max = 128) String newPassword
+) {}
