+++ auth/src/main/java/pt/estga/auth/oauth/GoogleTokenVerifier.java
@@ -1,4 +1,4 @@
-package pt.estga.config;
+package pt.estga.auth.oauth;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
