+++ auth/src/main/java/pt/estga/auth/oauth/OAuth2LoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package pt.estga.config;
+package pt.estga.auth.oauth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -8,6 +8,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
+import pt.estga.auth.config.JwtService;
 import pt.estga.user.entities.Role;
 import pt.estga.user.entities.User;
 import pt.estga.user.repositories.UserRepository;
