+++ pom.xml
@@ -100,6 +100,11 @@
             <!-- Internal modules -->
             <dependency>
                 <groupId>pt.estga</groupId>
+                <artifactId>auth</artifactId>
+                <version>${stonemark.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>pt.estga</groupId>
                 <artifactId>common-core</artifactId>
                 <version>${stonemark.version}</version>
             </dependency>
