+++ user/src/main/java/pt/estga/user/services/UserService.java
@@ -3,7 +3,6 @@ package pt.estga.user.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pt.estga.commoninfra.jpa.SpecBuilder;
@@ -24,7 +23,6 @@ public class UserService {
 
     private final UserRepository repository;
     private final ChatbotAccountRepository chatbotAccountRepository;
-    private final PasswordEncoder passwordEncoder;
 
     public Page<UserDto> search(UserFilter filter, Pageable pageable) {
         var sb = new SpecBuilder<User>()
@@ -91,32 +89,6 @@ public class UserService {
         User user = repository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
         UserMapper.update(user, request);
-        repository.save(user);
-    }
-
-    @Transactional
-    public void setPassword(Long userId, String rawPassword) {
-        User user = repository.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
-        if (user.getPasswordHash() != null) {
-            throw new IllegalArgumentException("Password is already set. Use PUT to change it.");
-        }
-        user.setPasswordHash(passwordEncoder.encode(rawPassword));
-        repository.save(user);
-    }
-
-    @Transactional
-    public void changePassword(Long userId, String oldPassword, String newPassword) {
-        User user = repository.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
-        if (user.getPasswordHash() == null) {
-            throw new IllegalArgumentException("No password set. Use POST to set one first.");
-        }
-        if (!passwordEncoder.matches(oldPassword, user.getPasswordHash())) {
-            throw new IllegalArgumentException("Current password is incorrect.");
-        }
-        user.setPasswordHash(passwordEncoder.encode(newPassword));
-        user.setTokenVersion(user.getTokenVersion() + 1);
         repository.save(user);
     }
 }
